infra/pyproject.toml
[tool.poetry]
name = "infra-eternal-dice"
version = "0.1.0"
description = ""
authors = ["lesteenman <lesteenman@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
"aws-cdk.core" = "^1.91.0"
"aws-cdk.aws-apigateway" = "^1.91.0"
"aws-cdk.aws-lambda" = "^1.91.0"
"aws-cdk.aws-logs-destinations" = "^1.91.0"
flake8 = "^3.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
